% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindMatches.R
\name{bindMatches}
\alias{bindMatches}
\title{bindMatches}
\usage{
bindMatches(dfA, dfB, inds, idA, idB)
}
\arguments{
\item{dfA}{First data frame to bind}

\item{dfB}{Second data frame to bind}

\item{inds}{A nx2 index pair matrix to bind on}

\item{idA}{What to concatenate to a column name in dfA if there exists a identical column in dfB}

\item{idB}{What to concatenate to a column name in dfB if there exists a identical column in dfB}
}
\value{
\code{bindMatches} returns a dataframe of the original dataframes bound by index.
}
\description{
Given 2 data frames(dfA and dfB) and a nx2 matrix of matched indices, will return a dataframe containing
both dataframes bound by index. If there are identical names in both dataframes, the result
will differentiate by adding idA(default is "A.") to the duplicate names in dfA, and idB(default is "B.") to those in dfB.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
