% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindMatches.R
\name{bindMatches_p}
\alias{bindMatches_p}
\title{bindMatches_p}
\usage{
bindMatches(dfA, dfB, inds, p, idA, idB)
}
\arguments{
\item{dfA}{First data frame to bind}

\item{dfB}{First data frame to bind}

\item{inds}{A nx2 index pair matrix to bind on}

\item{idA}{What to concatenate to a column name in dfA if there exists a identical column in dfB}

\item{idB}{What to concatenate to a column name in dfB if there exists a identical column in dfB}

\item{P}{a nx1 vector of the probability of each pair}
}
\value{
\code{bindMatches} returns a dataframe of the original dataframes bound by index.
}
\description{
Same as bindMatches, but also binds an extra column for the probability of each match
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
