% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corr.numeric}
\alias{corr.numeric}
\title{corr.numeric}
\usage{
corr.numeric(val, scale = 1)
}
\arguments{
\item{val}{The value to be corrupted. Should be numeric.}

\item{scale}{The scale applied to the error.}
}
\value{
The value  + the generated error.
}
\description{
A general numeric corruption function. Adds a value from the normal distribution (scaled by scale) to the value.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
