% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corr_replace}
\alias{corr_replace}
\title{corr_replace}
\usage{
corr_replace(val, pool = NULL)
}
\arguments{
\item{val}{The value to be corrupted. Should be string}

\item{scale}{The scale applied to the error.}
}
\value{
The value  + the generated error.
}
\description{
A general replacement corruption function. Takes a value and a pool of values, and returns a random element from the pool.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
