% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corr_string}
\alias{corr_string}
\title{corr_string}
\usage{
corr_string(val, scale = 2)
}
\arguments{
\item{val}{The value to be corrupted. Should be string}

\item{scale}{The scale applied to the error.}
}
\value{
The value  + the generated error.
}
\description{
A general string corruption function. Either adds, deletes, or substitutes single characters in the string.
The number of insertions, deletions, and replacements is generated from the normal, scaled by scale.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
