% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corrupt_character}
\alias{corrupt_character}
\title{corrupt_character}
\usage{
corrupt_character(vals, error_rate = 0.2, ...)
}
\arguments{
\item{error_rate}{The probability of the value being corrupted. Default is 2.}

\item{val}{The value to be corrupted. Should be string}

\item{scale}{The scale applied to the error.}
}
\value{
The value passed in, or a corrupted value with probability error rate.
}
\description{
A general string corruption function. Either adds, deletes, or substitutes single characters in the string.
The number of insertions, deletions, and replacements is generated from the normal, scaled by scale.
A built in corruption function, ready to be passed to corrupt data. Does not need to be passed to corrupt.factory.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
