% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corrupt_data}
\alias{corrupt_data}
\title{corrupt_data}
\usage{
corrupt_data(df, ..., .v = FALSE, col_names = NULL,
  col_contams = NULL)
}
\arguments{
\item{df}{Data frame to be corrupted}

\item{numeric_cor}{Names of columns to be corrupted numerically}

\item{string_cor}{Names of columns to be corrupted with string corruption.}

\item{replace_cor}{Names of columns to be corruped with replacement corruption}

\item{er}{Vector of error rates for each function, or single value to be applied by all functions. Default is .1.}

\item{v}{Verbose tag. Default is false}
}
\value{
Corrupted data frame
}
\description{
Takes data frame, returns data frame where each column is subjected to the passed in corruption functions.
}
\examples{
data_size = 500
rec <- generate_data(data_size)
error <-  .6
corrupt_rec <- corrupt_data(rec, numeric_cor = c("zipcode","birthyear"),string_cor = c("names"),replace_cor = c("town"),er = error, v = TRUE)


}
