% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corrupt_numeric}
\alias{corrupt_numeric}
\title{corrupt_numeric}
\usage{
corrupt_numeric(vals, error_rate = 0.2, ...)
}
\arguments{
\item{error_rate}{The probability of the value being corrupted. Default is 2.}

\item{val}{The value to be corrupted. Should be numeric.}

\item{scale}{The scale applied to the error. Default is 1.}
}
\value{
Either the value handed in, or the value  + the generated error with probability error_rate.
}
\description{
A general numeric corruption function. Adds a value from the normal distribution (scaled by scale) to the value.
A built in corruption function, ready to be passed to corrupt data. Does not need to be passed to corrupt.factory.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
