% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_data.R
\name{corrupt_replace}
\alias{corrupt_replace}
\title{corrupt_replace}
\usage{
corrupt_replace(vals, error_rate = 0.2, ...)
}
\arguments{
\item{error_rate}{The probability of the value being corrupted. Default is 2.}

\item{val}{The value to be corrupted. Should be string}

\item{pool}{A vector to pull the replacement values from.}
}
\value{
The value passed in, or a replaced value with probability error rate.
}
\description{
A general replacement corruption function. Takes a value and a pool of values, and returns a random element from the pool.
A built in corruption function, ready to be passed to corrupt data. Does not need to be passed to corrupt.factory.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
