% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_record_data.R
\name{generate_record_data}
\alias{generate_record_data}
\title{generate_record_data}
\usage{
generate_record_data(s)
}
\arguments{
\item{s}{The number of samples to generate}
}
\value{
A (s)x6 dataframe containing record information
}
\description{
A function that generates random data to use for testing record linkage packages.
The generated data contains name, birthyear, zipcode, town, sex, and id.
}
