% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfusion.R
\name{getConfusion}
\alias{getConfusion}
\title{getConfusion}
\usage{
getConfusion(dfA, dfB, inds, idColA, idColB = NULL)
}
\arguments{
\item{dfA}{First data frame being linked}

\item{dfB}{Second data frame being linked}

\item{inds}{A nx2 index pair matrix indicating the detected matches}

\item{idColA}{The column name in dfA of the shared id across dfA and dfB}

\item{idColB}{The column name in dfB of the shared id across dfA and dfB. If NULL, it assumes the column in B has the same name as the column in A}
}
\value{
\code{getConfusion} returns a dataframe of the original dataframes bound by index.
}
\description{
getConfusion
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
