% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfusion.R
\name{getConfusionInfo}
\alias{getConfusionInfo}
\title{getConfusionInfo}
\usage{
getConfusionInfo(conMat)
}
\arguments{
\item{conMat}{The confusion matrix generated by getConfusion}
}
\value{
\code{getConfusionInfo} Returns a dataframe of information on the given confusion table, including sensitivity, specificity, PPV, NPV, etc.
}
\description{
getConfusionInfo
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
