% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_record_data.R
\name{sample_names_uniform}
\alias{sample_names_uniform}
\title{sample_names_uniform}
\usage{
sample_names_uniform(s)
}
\arguments{
\item{s}{The number of samples to produce}
}
\value{
(s)x2 dataframe containing zipcode and town samples.
}
\description{
A function for generating records from the built in data included in this package.
This function generates \code{s}
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
