% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_record_data.R
\name{sample_towns_by_pop}
\alias{sample_towns_by_pop}
\title{sample_towns_by_pop}
\usage{
sample_towns_by_pop(s)
}
\arguments{
\item{s}{The number of samples to produce}
}
\value{
Character vector containing s towns
}
\description{
A function for generating records from the built in data included in this package.
This function generates \code{s} towns, such that the towns are chosen with frequency that
represents the total population of that town.
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
