% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_record_data.R
\name{sample_zips_by_pop}
\alias{sample_zips_by_pop}
\title{sample_zips_by_pop}
\usage{
sample_zips_by_pop(s)
}
\arguments{
\item{s}{The number of samples to produce}
}
\value{
(s)x2 dataframe containing zipcode and town samples.
}
\description{
A function for generating records from the built in data included in this package.
This function generates \code{s} zipcodes, such that the zipcodes are chosen with frequency that
represents the total population of that zipcode
}
\author{
Sam Murray<slmurray@andrew.cmu.edu>
}
