% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplyArgs.R
\name{supplyArgs}
\alias{supplyArgs}
\title{supplyArgs}
\usage{
supplyArgs(func, ...)
}
\arguments{
\item{func}{Any non-primative function with arguments.}

\item{...}{Name value pairs of arguments to func, and their associated value.}
}
\description{
Takes a nonprimitive function with arguments, and returns a modified function supplies default values for named arguments of the function.
Default values ARE overridable.
}
\details{
If multiple name value pairs in ... contain the same name, supplyArgs will through an error.
}
\author{
Sam Murray <slmurray@andrew.cmu.edu>
}
