% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplyArgs.R
\name{typeCheck}
\alias{typeCheck}
\title{typeCheck}
\usage{
checkArgs(func, ...)
}
\arguments{
\item{val}{Any value}

\item{check}{Either a string that identifies a class ("numeric"), a function identifies a class (is.numeric), or a list of strings and functions.}
}
\description{
Takes a type or type check function and a value and determines if that value is of that type.
}
\author{
Sam Murray <slmurray@andrew.cmu.edu>
}
