% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplyArgs.R
\name{wrapFunc}
\alias{wrapFunc}
\title{wrapFunc}
\usage{
wrapFunc(func, wrapFunc)
}
\arguments{
\item{func}{Any non-primative function with arguments.}

\item{wrapFunc}{Any function that can take as input the output of func. If this is not flexable enough, use supplyArgs or composeFunc.}
}
\description{
Takes two functions\code{(func, wrapFunc)}, returns the function composition of the two funtions \code{(wrapFunc(func()))}.
The return value will always have same arguments as func.
}
\details{
If multiple name value pairs in ... contain the same name, supplyArgs will through an error.
}
\author{
Sam Murray <slmurray@andrew.cmu.edu>
}
